Implemented both AddFolder and AddReceipt (#3)
Added functionality for adding folders. This can be done from the
home screen or within adding a receipt. It allows for giving a list
of business names associated with it, and accepts no values as valid
as it will just not recommend with ML but still works.

Updated DB to handle the cursor parsing to keep DB functionality
more contained and easy to use.

Adding some funky styling.